refactor: move sso/ and storage/ to src/providers directory
