+++ src/modules/authentications/authentications.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { KeycloakRolesService } from 'src/sso/keycloak/roles.provider';
+import { KeycloakRolesService } from '../../providers/sso/keycloak/roles.provider';
 import { AuthenticationsController } from './authentications.controller';
 import { AuthenticationsService } from './authentications.service';
