+++ src/modules/files/files.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { FilesController } from './files.controller';
-import { MinioClientModule } from './../../storage/minio/minio.module';
+import { MinioClientModule } from '../../providers/storage/minio/minio.module';
 import { FilesService } from './files.service';
 import { HttpModule } from '@nestjs/axios';
